Enhacements

+++ API/InventoryManagerAPI/DTO/InvoiceDTO.cs
@@ -6,15 +6,15 @@ namespace InventoryManagerAPI.DTO
     {
         public string InvoiceNumber { get; set; } = null!;
         public DateTime InvoiceDate { get; set; }
-        public long TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
 
-        public long Discount { get; set; }
+        public decimal Discount { get; set; }
         public string Status { get; set; } = null!;
         public string Comments { get; set; } = null!;
         public Customer customer { get; set; }
-        public long Profit { get; set; }
+        public decimal Profit { get; set; }
 
-        public long AmountPaid { get; set; }
+        public decimal AmountPaid { get; set; }
 
         public List<ProductDTO> productList { get; set; }
     }
