+++ API/InventoryManagerAPI/Models/Invoice.cs
@@ -13,14 +13,14 @@ namespace InventoryManagerAPI.Models
         public Guid InvoiceId { get; set; }
         public string InvoiceNumber { get; set; } = null!;
         public DateTime InvoiceDate { get; set; }
-        public long TotalAmount { get; set; }
-        public long Discount { get; set; }
+        public decimal TotalAmount { get; set; }
+        public decimal Discount { get; set; }
         public string Status { get; set; } = null!;
         public string Comments { get; set; } = null!;
         public Guid CustomerId { get; set; }
-        public long Profit { get; set; }
-        public long? AmountPaid { get; set; }
-        public long? AmountDue { get; set; }
+        public decimal Profit { get; set; }
+        public decimal? AmountPaid { get; set; }
+        public decimal? AmountDue { get; set; }
 
         public virtual Customer Customer { get; set; } = null!;
         public virtual ICollection<InvoiceProduct> InvoiceProducts { get; set; }
